requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
lockfile==0.12.2
pkg-resources==0.0.0
python-geoip==1.2
python-geoip-geolite2==2015.303
requests==2.22.0
urllib3==1.25.7
